extract apollo compatibility bits into a separate class (saves some bundle size)
